# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/prices": {
            "get": {
                "tags": [
                    "Цены"
                ],
                "summary": "Получить список всех цен",
                "operationId": "63348bf88df6d4eb68dfd5192429af28",
                "responses": {
                    "200": {
                        "description": "Список всех цен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Price"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Цены"
                ],
                "summary": "Создать новую цену",
                "operationId": "6a8846778186567e81a8ae51b36fabbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID цены",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "configuration_id": {
                                        "description": "ID конфигурации",
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "price": {
                                        "description": "цена",
                                        "type": "number",
                                        "format": "float",
                                        "example": "135000.50"
                                    },
                                    "start_date": {
                                        "description": "Дата начала лействия цены",
                                        "type": "string",
                                        "example": "2024-04-01T10:00:00Z"
                                    },
                                    "end_date": {
                                        "description": "Дата окончания действия цены (Если установлена)",
                                        "type": "string",
                                        "example": "2024-04-25T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданная цена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Price"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/prices/{id}": {
            "get": {
                "tags": [
                    "Цены"
                ],
                "summary": "Получить информацию о конкретной цене по ID",
                "operationId": "a573c52428ed8649d28f25cf8388580b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о цене",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Price"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Цена не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Цены"
                ],
                "summary": "Удалить цену",
                "operationId": "affc9c47028dfff617ab6f9da7280b26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Цена удалена"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Цены"
                ],
                "summary": "Обновить цену",
                "operationId": "a9efdec779d6fde8ea55216f7011ebf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цены",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID цены",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "configuration_id": {
                                        "description": "ID конфигурации",
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "price": {
                                        "description": "цена",
                                        "type": "number",
                                        "format": "float",
                                        "example": "135000.50"
                                    },
                                    "start_date": {
                                        "description": "Дата начала лействия цены",
                                        "type": "string",
                                        "example": "2024-04-01T10:00:00Z"
                                    },
                                    "end_date": {
                                        "description": "Дата окончания действия цены (Если установлена)",
                                        "type": "string",
                                        "example": "2024-04-25T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновлённая цена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Price"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Price": {
                "title": "Цена",
                "description": "Цена для комплектации автомобиля",
                "required": [
                    "configuration_id",
                    "price",
                    "start_date"
                ],
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор цены",
                        "type": "integer",
                        "example": 1
                    },
                    "configuration_id": {
                        "description": "ID комплектации, к которой относится цена",
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "description": "Цена комплектации",
                        "type": "number",
                        "format": "float",
                        "example": 35000.5
                    },
                    "start_date": {
                        "description": "Дата начала действия цены",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:00:00Z"
                    },
                    "end_date": {
                        "description": "Дата окончания действия цены (Если установлена)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T10:00:00Z"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}